pyproject.toml
[tool.poetry]
name = "bot-ui"
version = "0.1.0"
description = "This is bot's pl visualize ui"
authors = ["satellitex <s.a.t.e.3.ths@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
plotly = "^5.13.0"
dash = "^2.8.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
